Fix bug in mt_get_char
This function was returning dangling pointers in some cases.  The
multi_t tagged union argument was passed by value and in some cases a
pointer was returned that referenced the contents of this union.

This appears to be a rather serious bug as hash buckets appear to be
computed using the returned pointer.

This commit passes a pointer to the tagged union rather than passing it
by value.