Fix external symbols linkage
The existing __attribute__ ((common)) approach appears to only work if
everything is built in one run, which seems to prevent us from splitting
up the build at all.  For example, the current branch aims to share a
static library between the magmad executable and the test executable.
Doing so using the current approach results in multiple definition
errors.

This commit switches to a more traditional extern approach, which has
the disadvantage of duplicating symbols from the bundled dependencies in
both symbols.c and symbols.h.