Bug fix regarding all references to 'count' stringer:
    count  = uint24_put_no(...)

allocates a stringer on the heap which must be freed or you leak
heap memory.  Count was being treated as though it allocated
stringers on the stack and was not a concern.  I added error checks
and st_free calls to help manage all references to new stringers
created via uint24_put_no().

uint24_put_no() allocates a stringer via a call to st_alloc().
Precidence for how to handle a stringer returned from
st_alloc() is in bitwise.c. That shows a need to call st_free()
on the new stringer as if it lives on the heap when you're finished
using it.