Library build recursion fixed. Checks are once again run in parallel.
When not using the symlink, build.lib.sh would have problems calling
itself and passing the appropriate parameters to the child.

Checks are run in parallel but the other steps will stay sequential
for now. Some of the dependencies reference other libraries in the 
bundle which need to be configured and built first. Since those steps
run far quicker, and the difficulty in unraveling the connections is 
high, they are staying sequential... for now.