Bug fixes to allow magma to accept inbound mail over SMTP.
Removed a call to st_free() in smtp_rcpto() that was freeing a buffer
and then continuing to use the already freed memory.

Fixed the call to smtp_fetch_inbound() so it parses the domain portion
out before it tries searching for a match using domain_wildard(). Still
needs testing though.

Fixed the smtp_fetch_inbound() so it only stores the public key if 
the secure flag is set, and is now decoding the key before storing it.