Pipelined and line based SSL connections are finally working!
This means Thunderbird and the t.inbound (for example) are both working
at the same time. The issue was with SSL_peek blocking. Ultimately
I decided to have con_read simply return if data was already in the 
connection buffer, rather that see if additional data was available
on the wire before returning.